+++ src/app/app.module.ts
@@ -12,7 +12,7 @@ import { HomeComponent } from './views/home/home.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MaterialModule } from './shared/material.module';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { BicycleTypesComponent } from './components/bicycle-types/bicycle-types.component';
+import { BicycleTypesComponent } from './components/home/bicycle-types/bicycle-types.component';
 
 @NgModule({
   declarations: [
