--- src/app/components/bicycle-types/bicycle-types.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-bicycle-types',
-  templateUrl: './bicycle-types.component.html',
-  styleUrls: ['./bicycle-types.component.scss']
-})
-export class BicycleTypesComponent {
-
-}
