+++ src/app/components/home/about/about.component.html
@@ -6,7 +6,7 @@
           <div class="p-6 col-lg-5 mr-5">
             <div class="p-2">
               <h1 class="font-weight-bold">Lead Your Way</h1>
-              <p class="text-left text-weight-light h4 my-5 text-lg">
+              <p class="text-left text-weight-light fs-4 my-5">
                 Somos una compañía que ofrece una plataforma digital donde podrás encontrar bicicletas para todo tipo de usuarios. Por lo que, lo que
                 necesitas es una bicicleta tanto para practicar deporte o necesitas otro tipo de movilidad, en esta página encontrarás todo lo
                 necesario de forma rápida, sencilla y segura. Además, nunca tendrás que preocuparte por no encontrar la bicicleta adecuada, en este
