+++ src/app/components/home/bicycle-types/bicycle-types.component.html
@@ -1,0 +1,47 @@
+<section id="bicycle-types" class="bg-light m-0">
+  <div class="col-md-10 py-5 px-3 mx-auto text-center">
+    <h2 class="h3 my-5 text-weight-light">
+      <b>Lead Your Way</b> es la forma <b>más fácil</b><br />
+      de buscar bicicletas para...
+    </h2>
+    <div class="cotainer my-3">
+      <div class="row mx-auto">
+        <div class="col-md-4">
+          <a href="https://www.flaticon.es/iconos-gratis/ciclista" title="cilista iconos"
+            ><img src="/assets/images/road.png" alt="Road" width="80" />
+          </a>
+          <h4 class="">Ruta</h4>
+          <p class="my-3">
+            La bicicleta de ruta se utiliza para describir bicicletas construidas para viajar a gran velocidad en carreteras pavimentadas. Sus
+            características están pensadas para llevarte tan lejos y tan rápido como tus piernas lo permitan. Tienen un diseño con ángulos de asiento
+            y frontales muy verticales, un eje pedalier alto y muy poca curvatura de la horquilla.
+          </p>
+        </div>
+        <div class="col-md-4">
+          <a href="https://www.flaticon.es/iconos-gratis/montana" title="montana iconos">
+            <img src="/assets/images/mountain.png" alt="Mountain" width="80" />
+          </a>
+          <h4 class="mt-3">Montaña</h4>
+          <p class="my-3">
+            Son bicicletas que, o bien carecen de suspensión o tienen un sistema de amortiguación delantera en la horquilla. Normalmente el recorrido
+            de esta suspensión es de menos de 120mm, ya que están pensadas para rodar a toda velocidad por pistas y tramos sin asfaltar sin grandes
+            desniveles ni obstáculos pronunciados.
+          </p>
+        </div>
+        <div class="col-md-4">
+          <a href="https://www.flaticon.es/iconos-gratis/city" title="city icons">
+            <img width="80" src="/assets/images/city.png" alt="City" />
+          </a>
+          <h4 class="mt-3">Ciudad</h4>
+          <p class="my-3">
+            Una bicicleta urbana se caracteriza por tener componentes simples pero resistentes a cualquier picadura. Su mantenimiento es más sencillo
+            y por lo tanto económico. En comparación con una bicicleta de montaña suelen ser más pesadas y no requieres suspensiones para los
+            recorridos en la ciudad.
+          </p>
+          <p></p>
+        </div>
+      </div>
+    </div>
+    <a class="btn btn-primary text-weight-bold" href="#"> Buscar bibicletas </a>
+  </div>
+</section>
