+++ src/app/components/home/bicycle-types/bicycle-types.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-bicycle-types',
+  templateUrl: './bicycle-types.component.html',
+  styleUrls: ['./bicycle-types.component.scss']
+})
+export class BicycleTypesComponent {
+
+}
