+++ src/app/components/home/contact/contact.component.html
@@ -2,7 +2,7 @@
   <div class="container contact-form">
     <div class="contact-image">
       <a href="https://www.flaticon.es/iconos-gratis/cohete" title="cohete iconos">
-        <img src="imgs/rocket.png" alt="rocket_contact" />
+        <img src="assets/images/rocket.png" alt="rocket_contact" />
       </a>
     </div>
     <h3 class="my-4 text-weight-bold">Déjanos un mensaje</h3>
